[Feat] #23  피드 상세 화면 API 연동

+++ app/src/main/java/com/project/score/API/response/user/FeedListResponse.kt
@@ -30,7 +30,7 @@ data class PageableInfo(
 )
 
 data class FeedListResponse(
-    val feedId: Long,
+    val feedId: Int,
     val uploaderNickname: String,
     val uploaderProfileImgUrl: String,
     val feedImg: String,
